扩展模块增删改

+++ uoo-core/uoo-public/src/main/java/cn/ffcs/uoo/core/expando/controller/TbExpandovalueController.java
@@ -2,9 +2,9 @@ package cn.ffcs.uoo.core.expando.controller;
 
 
 import cn.ffcs.uoo.base.common.annotion.UooLog;
-import cn.ffcs.uoo.base.common.tool.util.DateUtils;
+import cn.ffcs.uoo.base.common.tool.util.StringUtils;
 import cn.ffcs.uoo.base.controller.BaseController;
-import cn.ffcs.uoo.core.dictionary.vo.ResponseResult;
+import cn.ffcs.uoo.core.vo.ResponseResult;
 import cn.ffcs.uoo.core.expando.entity.TbExpandorow;
 import cn.ffcs.uoo.core.expando.entity.TbExpandovalue;
 import cn.ffcs.uoo.core.expando.service.TbExpandorowService;
@@ -14,10 +14,13 @@ import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 /**
  * <p>
@@ -96,14 +99,25 @@ public class TbExpandovalueController extends BaseController {
             return responseResult;
         }
 
+        TbExpandovalue tbExpandovalue = new TbExpandovalue();
+        tbExpandovalue.setValueId(valueId);
+        tbExpandovalue.setStatusCd("1000");
+
+        List<TbExpandovalue>  tbExpandovalueList = tbExpandovalueService.selectValueList(tbExpandovalue);
+        if(tbExpandovalueList == null || tbExpandovalueList.size() <= 0) {
+            responseResult.setState(ResponseResult.STATE_ERROR);
+            responseResult.setMessage("该扩展值id对应的扩展值不存在");
+            return responseResult;
+        }
+
         // 查询出对应的扩展行id
-        TbExpandovalue tbExpandovalue = tbExpandovalueService.selectById(valueId);
+        TbExpandovalue tbExpandovalueOut = tbExpandovalueList.get(0);
 
         // 删除扩展值
         tbExpandovalueService.remove(valueId, updateUser);
 
         // 删除扩展行
-        tbExpandorowService.remove(tbExpandovalue.getRowId(), updateUser);
+        tbExpandorowService.remove(tbExpandovalueOut.getRowId(), updateUser);
 
         responseResult.setState(ResponseResult.STATE_OK);
         responseResult.setMessage("删除成功");
@@ -123,27 +137,46 @@ public class TbExpandovalueController extends BaseController {
             responseResult.setMessage("请输入扩展值标识");
             return responseResult;
         }
-        if (tbExpandovalue.getTableId() == null) {
-            responseResult.setState(ResponseResult.STATE_ERROR);
-            responseResult.setMessage("请输入系统表标识");
-            return responseResult;
-        }
-        if (tbExpandovalue.getColumnId() == null) {
-            responseResult.setState(ResponseResult.STATE_ERROR);
-            responseResult.setMessage("请输入扩展列标识");
-            return responseResult;
-        }
-        if (tbExpandovalue.getRowId() == null) {
-            responseResult.setState(ResponseResult.STATE_ERROR);
-            responseResult.setMessage("请输入扩展行标识");
-            return responseResult;
-        }
 
         tbExpandovalueService.updateById(tbExpandovalue);
 
         responseResult.setState(ResponseResult.STATE_OK);
         responseResult.setMessage("修改扩展值成功");
         return responseResult;
+    }
+
+    @ApiOperation(value = "查询扩展值列表", notes = "查询扩展值列表")
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "resourceId", value = "资源标识", required = true, dataType = "String", paramType = "path"),
+            @ApiImplicitParam(name = "tableId", value = "系统表标识", required = true, dataType = "Long", paramType = "path"),
+            @ApiImplicitParam(name = "columnId", value = "扩展列标识", required = true, dataType = "Long", paramType = "path"),
+            @ApiImplicitParam(name = "recordId", value = "业务记录标识", required = true, dataType = "Long", paramType = "path")
+    })
+    @UooLog(value = "查询扩展值列表", key = "queryValueList")
+    @RequestMapping(value = "/getList/{resourceId}/{tableId}/{columnId}/{recordId}", method = RequestMethod.GET)
+    public List<TbExpandovalue> queryValueList(@PathVariable String resourceId, @PathVariable Long tableId,
+                                               @PathVariable Long columnId, @PathVariable String recordId) {
+        // 校验必填项
+        if (StringUtils.isEmpty(resourceId)) {
+            return null;
+        }
+        if (tableId == null) {
+            return null;
+        }
+        if (columnId == null) {
+            return null;
+        }
+        if (StringUtils.isEmpty(recordId)) {
+            return null;
+        }
+
+        TbExpandovalue tbExpandovalue = new TbExpandovalue();
+        tbExpandovalue.setResourceId(resourceId);
+        tbExpandovalue.setTableId(tableId);
+        tbExpandovalue.setColumnId(columnId);
+        tbExpandovalue.setRecordId(recordId);
+
+        return tbExpandovalueService.selectValueList(tbExpandovalue);
     }
 }
